# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Convoy API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List trips for the current user
      responses:
        "200":
          description: Trips list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TripSummary"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a trip (creator becomes leader)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripDetail"
  /trips/{tripId}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get trip details with roster
      parameters:
        - $ref: "#/components/parameters/TripId"
      responses:
        "200":
          description: Trip detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripDetail"
    patch:
      operationId: updateTrip
      tags: [trips]
      summary: Update trip (leaders only, Draft only for destination/route)
      parameters:
        - $ref: "#/components/parameters/TripId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripUpdate"
      responses:
        "200":
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripDetail"
  /trips/{tripId}/start:
    post:
      operationId: startTrip
      tags: [trips]
      summary: Start the trip (leader only)
      parameters:
        - $ref: "#/components/parameters/TripId"
      responses:
        "200":
          description: Trip started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripDetail"
  /trips/{tripId}/end:
    post:
      operationId: endTrip
      tags: [trips]
      summary: End the trip (leader only)
      parameters:
        - $ref: "#/components/parameters/TripId"
      responses:
        "200":
          description: Trip ended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripDetail"
  /trips/{tripId}/leave:
    post:
      operationId: leaveTrip
      tags: [members]
      summary: Leave the trip (blocked for the only leader)
      parameters:
        - $ref: "#/components/parameters/TripId"
      responses:
        "200":
          description: Left trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "409":
          description: Only leader cannot leave
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /join/{code}:
    get:
      operationId: getJoinPreview
      tags: [members]
      summary: Preview a trip by join code (no membership required)
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trip preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinPreview"
        "404":
          description: Unknown code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /join:
    post:
      operationId: joinTrip
      tags: [members]
      summary: Join a trip with a join code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinInput"
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripDetail"
  /trips/{tripId}/members:
    post:
      operationId: inviteMember
      tags: [members]
      summary: Invite a member by phone (leader only)
      parameters:
        - $ref: "#/components/parameters/TripId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInvite"
      responses:
        "201":
          description: Invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /trips/{tripId}/members/{memberId}:
    delete:
      operationId: removeMember
      tags: [members]
      summary: Remove a member (leader only)
      parameters:
        - $ref: "#/components/parameters/TripId"
        - $ref: "#/components/parameters/MemberId"
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "409":
          description: Cannot remove the only leader
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trips/{tripId}/members/{memberId}/promote:
    post:
      operationId: promoteMember
      tags: [members]
      summary: Promote a member to leader (leader only)
      parameters:
        - $ref: "#/components/parameters/TripId"
        - $ref: "#/components/parameters/MemberId"
      responses:
        "200":
          description: Promoted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
...
components:
  parameters:
    TripId:
      name: tripId
      in: path
      required: true
      schema:
        type: number
    MemberId:
      name: memberId
      in: path
      required: true
      schema:
        type: number
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    Coordinate:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
        lng:
          type: number
...
    TripInput:
      type: object
      required:
        [
          name,
          startLat,
          startLng,
          startLabel,
          destLat,
          destLng,
          destLabel,
          scheduledAt,
          routeGeometry,
          routeDistanceM,
          routeDurationS,
        ]
      properties:
        name:
          type: string
          minLength: 1
        startLat:
          type: number
        startLng:
          type: number
        startLabel:
          type: string
        destLat:
          type: number
        destLng:
          type: number
        destLabel:
          type: string
        scheduledAt:
          type: string
        routeGeometry:
          type: array
          items:
            $ref: "#/components/schemas/Coordinate"
        routeDistanceM:
          type: number
        routeDurationS:
          type: number
    TripUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        destLat:
          type: number
        destLng:
          type: number
        destLabel:
          type: string
        scheduledAt:
          type: string
        routeGeometry:
          type: array
          items:
            $ref: "#/components/schemas/Coordinate"
        routeDistanceM:
          type: number
        routeDurationS:
          type: number
    Trip:
      type: object
      required:
        [
          id,
          name,
          status,
          startLat,
          startLng,
          startLabel,
          destLat,
          destLng,
          destLabel,
          scheduledAt,
          routeDistanceM,
          routeDurationS,
          joinCode,
        ]
      properties:
        id:
          type: number
        name:
          type: string
        status:
          type: string
          enum: [draft, active, ended]
        startLat:
          type: number
        startLng:
          type: number
        startLabel:
          type: string
        destLat:
          type: number
        destLng:
          type: number
        destLabel:
          type: string
        scheduledAt:
          type: string
        routeDistanceM:
          type: number
        routeDurationS:
          type: number
        joinCode:
          type: string
        startedAt:
          type: ["string", "null"]
        endedAt:
          type: ["string", "null"]
    TripSummary:
      type: object
      required: [trip, memberCount, sharingCount, myRole]
      properties:
        trip:
          $ref: "#/components/schemas/Trip"
        memberCount:
          type: number
        sharingCount:
          type: number
        myRole:
          type: string
          enum: [leader, member]
    Member:
      type: object
      required: [id, name, phone, role, color, initial, joinStatus]
      properties:
        id:
          type: number
        userId:
          type: ["number", "null"]
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [leader, member]
        color:
          type: string
        initial:
          type: string
        vehicleType:
          type: ["string", "null"]
        joinStatus:
          type: string
          enum: [joined, invited]
        isSelf:
          type: boolean
    TripDetail:
      type: object
      required: [trip, members, routeGeometry, myMemberId, myRole]
      properties:
        trip:
          $ref: "#/components/schemas/Trip"
        members:
          type: array
          items:
            $ref: "#/components/schemas/Member"
        routeGeometry:
          type: array
          items:
            $ref: "#/components/schemas/Coordinate"
        myMemberId:
          type: number
        myRole:
          type: string
          enum: [leader, member]
    JoinPreview:
      type: object
      required: [tripName, destLabel, scheduledAt, status, memberCount]
      properties:
        tripName:
          type: string
        destLabel:
          type: string
        scheduledAt:
          type: string
        status:
          type: string
          enum: [draft, active, ended]
        memberCount:
          type: number
    JoinInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
          minLength: 1
    MemberInvite:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
          minLength: 5
        name:
          type: string
...