# lib/api-zod/src/generated/types/memberInvite.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Convoy API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MemberInvite {
  /** @minLength 5 */
  phone: string;
  name?: string;
}